social_media/requirements.txt
Django==2.2.4
django-crispy-forms==1.7.2
djangorestframework==3.10.3
pytz==2019.2
Pillow==7.0.0
dj-database-url==0.5.0
django-extensions==2.2.8
user-agents==2.1
django_user_agents
python-social-auth
social-auth-app-django
# for test
coverage